Move factory into separate file for clarity & scalability
